move invite and request rows
