+++ shared/teams/team/invites/invite-row/index.js
@@ -1,8 +1,8 @@
 // @flow
 import * as React from 'react'
-import {Avatar, Box, Button, ClickableBox, Text, Usernames} from '../../../common-adapters'
-import {globalStyles, globalMargins, isMobile} from '../../../styles'
-import {typeToLabel} from '../../../constants/teams'
+import {Avatar, Box, Button, ClickableBox, Text, Usernames} from '../../../../common-adapters'
+import {globalStyles, globalMargins, isMobile} from '../../../../styles'
+import {typeToLabel} from '../../../../constants/teams'
 
 export type Props = {
   label: string, // email, sbs, or seitan
